+++ src/misc/num_player_setter.cpp
@@ -27,7 +27,7 @@ void NumPlayerSetter::Init()
 	//initialize render slot locations
 	for(size_t it = 0; it < num_players_boxes.size(); it++)
 	{
-		num_players_boxes[it].render_rect = {320 - 25, 100 + 30*(it+1), 100, 20};
+		num_players_boxes[it].render_rect = {320 - 25, static_cast<float>( 100 + 30*(it+1) ), 100, 20};
 		if(it == 0)
 		{
 			num_players_boxes[it].text = " " + std::to_string( it + 1) + " Player";
